+++ src/qt_window.h
@@ -31,7 +31,7 @@ public:
     virtual void DrawBezierPath(const std::list<Point>& path, const Color color);
     virtual void DrawWavyLine(const int x1, const int y1, const int width, const int radius, const Color color);
     virtual void DrawImage(const int x1, const int y1, const int width, const int height, const std::vector<unsigned char>& bmp);
-    virtual float GetSymbolSize(const char32_t symbol, const int height, const std::string& family_name, Size& size, int& baseline);
+    virtual int GetSymbolSize(const char32_t symbol, const int height, const std::string& family_name, Size& size, int& baseline);
 
     virtual void ClearRect(const int x1, const int y1, const int width, const int height);
